Move from .zshrc to .zshrc.local
Nowadays, grml's zsh setup lives in only *one* zshrc file.
That is the global one: /etc/zsh/zshrc (from grml-etc-core).
It is best to leave *this* file untouched and do personal changes to
your zsh setup via ${HOME}/.zshrc.local which is loaded at the end of
the global zshrc.